requirements.txt
streamlit
google-generativeai
opencv-python
pillow
python-dotenv
fastapi
uvicorn 